+++ ckanext/harvest/cli.py
@@ -70,7 +70,7 @@ def show(ctx, id):
     try:
         with flask_app.test_request_context():
             result = utils.show_harvest_source(id)
-    except tk.ObjectNotFound as e:
+    except tk.ObjectNotFound:
         tk.error_shout(u"Source <{}> not found.".format(id))
         raise click.Abort()
     click.echo(result)
@@ -191,10 +191,38 @@ def job_abort(ctx, id):
     with flask_app.test_request_context():
         try:
             result = utils.abort_job(id)
-        except tk.ObjectNotFound as e:
+        except tk.ObjectNotFound:
             tk.error_shout(u"Job not found.")
             ctx.abort()
 
+    click.echo(result)
+
+
+@harvester.command()
+@click.argument("life_span", default=False, required=False)
+@click.option(
+    "-i",
+    "--include",
+    default=False,
+    help="""If source_id provided as included, then only it's failed jobs will be aborted.
+    You can use comma as a separator to provide multiple source_id's""",
+)
+@click.option(
+    "-e",
+    "--exclude",
+    default=False,
+    help="""If source_id provided as excluded, all sources failed jobs, except for that
+    will be aborted. You can use comma as a separator to provide multiple source_id's""",
+)
+@click.pass_context
+def abort_failed_jobs(ctx, life_span, include, exclude):
+    """Abort all jobs which are in a "limbo state" where the job has
+    run with errors but the harvester run command will not mark it
+    as finished, and therefore you cannot run another job.
+    """
+    flask_app = ctx.meta["flask_app"]
+    with flask_app.test_request_context():
+        result = utils.abort_failed_jobs(life_span, include, exclude)
     click.echo(result)
 
 
@@ -239,7 +267,8 @@ def run(ctx):
 @harvester.command()
 @click.pass_context
 @click.argument("id", metavar="SOURCE_ID_OR_NAME")
-def run_test(ctx, id):
+@click.argument("force-import", required=False, metavar="GUID")
+def run_test(ctx, id, force_import=None):
     """Runs a harvest - for testing only.
 
     This does all the stages of the harvest (creates job, gather,
@@ -248,9 +277,11 @@ def run_test(ctx, id):
     fire up gather/fetch_consumer processes, as is done in production.
 
     """
+    if force_import:
+        force_import_val = force_import.split('=')[-1]
     flask_app = ctx.meta["flask_app"]
     with flask_app.test_request_context():
-        utils.run_test_harvester(id)
+        utils.run_test_harvester(id, force_import_val)
 
 
 @harvester.command("import")
@@ -318,7 +349,7 @@ def import_stage(
                 package_id,
                 segments,
             )
-        except tk.ObjectNotFound as e:
+        except tk.ObjectNotFound:
             tk.error_shout(u"Source <{}> not found.".format(id))